chatbot/requirements.txt
streamlit
pdfplumber
pytesseract
pdf2image
Pillow
pandas>=2.2.0
langchain
openai
transformers==4.36.2
torch==2.2.1
numpy>=1.26.0
PyPDF2==3.0.1
scikit-learn==1.3.2
tqdm==4.66.1
datasets==2.15.0
